renames typography files
